services/ai_processor/requirements.txt
flask==2.3.3
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.0
redis==5.2.1
python-dotenv==1.0.0
gunicorn==21.2.0 